+++ Commandes/moderation/say.js
@@ -2,15 +2,9 @@ const Discord = require('discord.js')
 const config = require('../../config.json');
 module.exports.run = async (client, message, args) => {
     
-    if(!message.guild.me.hasPermission("MANAGE_MESSAGES")) {
-    const embed = new Discord.MessageEmbed()
-        .setColor(config.color)
-        .setDescription(message.language.errors.missingPerms(["MANAGE_MESSAGES"]))
-    return message.channel.send(embed);
-    }
     if(!message.member.hasPermission("MANAGE_MESSAGES")) {
         const embed = new Discord.MessageEmbed()
-            .setColor(config.color)
+            .setColor(client.color)
             .setDescription(message.language.errors.permLevel("MANAGE_MESSAGES"))
         return message.channel.send(embed);
     }
@@ -21,8 +15,11 @@ module.exports.run = async (client, message, args) => {
     }
 
     let messageToBot = args.join(' ')
-    message.delete().catch()
     message.channel.send(messageToBot)
+
+    if(message.guild.me.hasPermission("MANAGE_MESSAGES")) {
+        message.delete().catch()
+    }
 
 }
